added ExceptionBean, UncheckedException and CheckedException inherit from AbstractException now, deleted try/catch clauses in controllers

+++ service/src/main/java/ch/uzh/ifi/seal/soprafs15/service/exceptions/UserNotFoundException.java
@@ -1,6 +1,7 @@
 package ch.uzh.ifi.seal.soprafs15.service.exceptions;
 
 import org.springframework.http.HttpStatus;
+import org.springframework.web.bind.annotation.ResponseStatus;
 
 /**
  * Created by Hakuna on 13.04.2015.
@@ -14,7 +15,7 @@ public class UserNotFoundException extends UncheckedException {
         super(message, invoker, HttpStatus.NOT_FOUND);
     }
 
-    public UserNotFoundException(Integer id, Class invoker) {
+    public UserNotFoundException(Long id, Class invoker) {
         this("The user with id " + id + " couldn't be found.", invoker);
     }
 }
